+++ app/controllers/claimToAdjustPoa/CheckYourAnswersController.scala
@@ -28,7 +28,7 @@ import play.api.mvc._
 import services.claimToAdjustPoa.RecalculatePoaHelper
 import services.{ClaimToAdjustPoaCalculationService, ClaimToAdjustService, PaymentOnAccountSessionService}
 import uk.gov.hmrc.auth.core.AuthorisedFunctions
-import utils.ClaimToAdjust.{ClaimToAdjustUtils, WithSessionAndPoa}
+import utils.claimToAdjust.{ClaimToAdjustUtils, WithSessionAndPoa}
 import utils.AuthenticatorPredicate
 import views.html.claimToAdjustPoa.CheckYourAnswers
