+++ app/utils/claimToAdjust/JourneyCheckerClaimToAdjust.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package utils.ClaimToAdjust
+package utils.claimToAdjust
 
 import auth.MtdItUser
 import config.{AgentItvcErrorHandler, ItvcErrorHandler, ShowInternalServerError}
